requirements.txt
appnope==0.1.2
argcomplete==1.12.2
attrs==20.3.0
backcall==0.2.0
certifi==2020.12.5
commonroad-vehicle-models==2.0.0
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
derivative==0.3.1
importlib-metadata==3.7.3
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.36.0
matplotlib==3.3.4
more-itertools==8.7.0
numba==0.53.0
numpy==1.20.1
opencv-python==4.5.1.48
packaging==20.9
pandas==1.2.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.1
prompt-toolkit==3.0.17
ptyprocess==0.7.0
py==1.10.0
pygame==2.0.0.dev10
Pygments==2.8.1
pyparsing==2.4.7
pysindy==1.3.0
pytest==5.4.3
python-dateutil==2.8.1
pytz==2021.1
rope==0.18.0
scikit-learn==0.24.1
scipy==1.6.1
Shapely==1.7.1
six==1.15.0
sklearn==0.0
svgpathtools==1.4.1
svgwrite==1.4.1
threadpoolctl==2.1.0
torch==1.8.0
traitlets==5.0.5
typing-extensions==3.7.4.3
UPnPy==1.1.8
wcwidth==0.2.5
zipp==3.4.1
